requirements.txt
torch
torchvision
torchaudio
sentence-transformers
pinecone-client
flask
flask-cors
peft
bitsandbytes
langchain
faiss-gpu
